Do not just extend/include Resque::Helpers module
This module will be deprecated in the upcoming version 2.0 of resque and already posts an annyoing warning because of that. So we copy the methods used by resque-loner in its Helpers module.

In the future Resque.encode and Resque.decode will provide encoding/decoding facilities for JSON, so we check if this is already the case:
 - If yes, just use these methods,
 - otherwise delegate to the Resque::Helpers methods with some black magic while silencing the warning. This way we can use the new Resque-API even if it isn't all there yet.